The clipboard manager now checks if a service is available when activating or deactivating
The clipboard manager didn't check if a service is
available. This caused the app to crash if no service
was defined. This was caused by it tried to activate
the service itself which was not defined at the time.
The clipboard manager now saves the state locally
instead of asking the service.