Replaced the application lock system which fixes issue #79
The previous lock mechanism didn't work as well
because it just used a lock file. The app would
still think its running if the JVM was terminated.
This was because the lock file woudn't be removed.
This is now fixed by running a little local server
that listeners for connections if the app is running.
The app then checks if the app is running by checking
if it can connect to the server using the port specified
in the new lock file.