Loging in from the welcome activity no longer closes the app
At first the app closed when logging in from
the welcome activity. This was because the
main activity closed itself when opening the
welcome activity to make sure the user can't
go back to the main activity. This caused the
back stack to not contain the main activity
and thus closing the app. This has been fixed
by first opening the main activity without a
effect and then opening login activity so the
main activity will be in the back stack when
loggin in.