Moved the content pulled and pushed notifications to the messages file which resolves issue #85
I moved the content pulled and pushed messages
to the messages that is specific to the desktop.
The Android strings file that is being used right
now escapes quotes which are not parsed on the desktop
side. The best way so solve those issues is by moving
all the messages to the messages file. This also
allows for easier translations on the desktop side.