Added a history to the Android client which resolves issue #76
Added a history in place of the login button
on the dashboard. The history adds a entry to
the list when something is pulled, pushed,
share to another app and shared from another
app. It contains the content that was used,
the date and it also uses the dashboard icons
to show the user the action that it relates to.

In order to make this I had to use a SQLite
database and created a adapter for the history
list. In order to log the events to the history
I've added database inserts into the pull, push,
share to app and receive from app tasks.