Improved the app performance by using a executor for background tasks
Instead of creating a new thread every time
something needs to happen in the background the
clipboard and sync services now use a executor.
In the app code a executor gets set that will be
used by all the sub services. I also moved some
non UI tasks to this executor to limit UI lockups
and slowdowns.