Added a direct content getting and setting method to the clipboard service
I've added a method that waits till the clipboard content
is received or set before it returns. This allows easier
getting of the clipboard content because listening for
events and implementing a Clipboard Listener is nolonger
needed. In order to achieve this I've renamed the getContent
and setContent to requestGet and requestSet.