Moved the AWT clipboard service to the desktop project
I've moved the AWT clipboard service to the desktop project
to stop the Android build from failing

This service first was in the global project but because
Android doesn't have AWT it caused the Android build to
fail.
I've fixed this by moving it to the desktop project and
adding the service when the app starts in the app starting
class and not in the shell ui itself. Because of this it is
only used by the desktop.