The app now writes a .desktop file and icons to the filesystem
In order for Ubuntu to show a icon in the sync menu there has to be
a .desktop file describing the application. This also requires the
appropriate icons to be available.

This commit adds support for writing .desktop files and makes sure
that the app writes the 16x16 and 512x512 image to the filesystem.