Renamed the Settings class to Config
A config should be read on start and writen on shutdown. It shoudn't
be a place to enable a specific feature.

At the moment polling and other features are enabled by changing a
config value. This isn't a great way of doing it. This is a start
in improving the way of saving settings.