Added a catch to show a error when no SWT browser engine is available
Sometimes there is no engine available for the SWT browser. For
example on Ubuntu you can install 'libwebkitgtk-1.0-0' (at the time
of commiting) to fix this issue. This catch makes sure that the app
doesn't just crash but shows a nice message telling the user what
went wrong.

This fixes issue #32