Changed 'paste' to 'put in clipboard' in the queue
Just using 'paste' in the context menu is not very clear. It even
is the opposite of what actually does. This change makes it a bit
more clear to the user what the thing actually is about..

In order to do this I've added it the messages file under the key
'queue.context_menu.put_in_clipboard'.

This also resolves #38.