Added a UI thread check to the SWT clipboard service
I've added a check to the SWT clipboard service to see if calling
thread is the UI thread. If it is it will simply run the SWT
methods from the current thread. If not it will do it on the UI
thread. Because there is only one UI thread it can only execute one
thing at a time. This caused a deadlock because the SWT thread
was waiting on itself to finish.